added modal for delete button

+++ src/components/Results.js
@@ -5,24 +5,31 @@ import { Modal, Button } from 'react-bootstrap';
 function Results(props) {
   const [data, setData] = useState('');
   const [deleted, setDeleted] = useState(false);
+  const [loggedOut, setLoggedOut] = useState(false);
   const model = props.match.params.model;
   const year = props.match.params.year;
+  const make = props.match.params.make;
+  const loggedIn = props.loggedIn;
 
   const handleDelete = e => {
     e.preventDefault();
-    let id = e.target.value;
-    fetch(`https://total-garage.herokuapp.com/garage/repairs/${id}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `JWT ${localStorage.getItem('token')}`
-      },
-      mode: 'cors'
-    }).then(
-      setTimeout(() => {
-        setDeleted(true);
-      }, 125)
-    );
+    if (loggedIn) {
+      let id = e.target.value;
+      fetch(`https://total-garage.herokuapp.com/garage/repairs/${id}`, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `JWT ${localStorage.getItem('token')}`
+        },
+        mode: 'cors'
+      }).then(
+        setTimeout(() => {
+          setDeleted(true);
+        }, 125)
+      );
+    } else {
+      setLoggedOut(true);
+    }
   };
 
   useEffect(() => {
@@ -36,6 +43,36 @@ function Results(props) {
       )
       .then(setData);
   }, [deleted]);
+
+  const handleClose = () => setLoggedOut(false);
+
+  if (loggedOut) {
+    return (
+      <Modal.Dialog>
+        <Modal.Header>
+          <Modal.Title>Access Denied</Modal.Title>
+        </Modal.Header>
+
+        <Modal.Body>
+          <p>You have to be logged in to do that.</p>
+        </Modal.Body>
+
+        <Modal.Footer>
+          <a href="/signup">
+            <Button variant="warning">Signup</Button>
+          </a>
+
+          <a href="/login">
+            <Button variant="primary">Login</Button>
+          </a>
+
+          <Button variant="danger" onClick={handleClose}>
+            Close
+          </Button>
+        </Modal.Footer>
+      </Modal.Dialog>
+    );
+  }
 
   if (data) {
     return (
